+++ tests/spread/test_storage_detaching.py/task.yaml
@@ -1,0 +1,13 @@
+summary: test_storage_detaching.py
+environment:
+  TEST_MODULE: test_storage_detaching.py
+  # #1550 only reproduces on Juju 4.0: 3.6 masks it via a cleanup_storage shortcut
+  # that removes still-Dying storage. Pin this task to a 4.0 variant and drop the
+  # default juju36 variant so it runs ONLY on 4.0 (the rest of the suite stays 3.6).
+  CONCIERGE_JUJU_CHANNEL/juju40: 4.0/stable
+variants:
+  - -juju36
+execute: |
+  tox run -e integration -- "tests/integration/$TEST_MODULE" --model testing --alluredir="$SPREAD_TASK/allure-results"
+artifacts:
+  - allure-results
